Update inline documentation & settings for site branding
First, update the format of the file header comment to follow a style in keeping with JsDoc/PHPDoc for better consistency across the various types of programming work produced by the author for DAESA. Next, change the license from ISC to MIT, which is now preferred. Finally, update the overall branding of the SURCA website to better match the branding of the DAESA website.