Styling: Make enhancements to CSS/wsuwp-customizer_addit…
Make the following style enhancements in CSS/wsuwp-customizer_additional-css.css for the abstract finder tool: 1) Modify the layout and styling of abstract cards so that they look different when receiving focus and are able to cover other components on the page without visual clashing. 2) Design a layout for effecting a modal viewing mode for project cards which the user can trigger. 3) Create a style for a modal cover that serves as a background layer below a project card opened in a modal viewing mode.