Production: Update HTML/home.html
Update markup in HTML/home.html used to maintain the SURCA home page via the WSUWP Spine Theme's builder template with the following changes: 1) Update the countdown timer's expiration message to announce the awards ceremony on March 30. 2) Update the cards on the home page to link to the awards ceremony watch page and display Bev's recent news article discussing how virtual SURCA 2021 will be conducted. 3) Make the invitation more specific to the audiences we are inviting.